Integrated privacy policy tests with rest of test suite
