+++ symphoni-ui/Tests/test_suite.py
@@ -3,6 +3,7 @@ import unittest
 import join_button
 import start_party_button
 import queue
+import privacy_button
 
 loader = unittest.TestLoader()
 suite = unittest.TestSuite()
@@ -11,5 +12,6 @@ runner = unittest.TextTestRunner(verbosity=3)
 suite.addTests(loader.loadTestsFromModule(join_button))
 suite.addTests(loader.loadTestsFromModule(start_party_button))
 suite.addTests(loader.loadTestsFromModule(queue))
+suite.addTests(loader.loadTestsFromModule(privacy_button))
 
 runner.run(suite)
