+++ symphoni-ui/public/index.html
@@ -31,7 +31,7 @@
     <h3 class="center">
       <a class="waves-effect waves-light btn-large z-depth-3" href="#start" id="start-button">Start Party!</a>
       <a class=" waves-effect waves-light btn-large z-depth-3 " href="#join" id="join-button">Join Party</a>
-      <a class="waves-effect waves-light btn-large" href="#privacy">Privacy Policy</a>
+      <a class="waves-effect waves-light btn-large" href="#privacy" id="privacy-button" >Privacy Policy</a>
     </h3>
 
     <!-- START PARTY CONTAINER -->
