--- symphoni-ui/ui-test/requirements.txt
@@ -1,1 +1,0 @@
-selenium==3.14.1
